Create README.md
I am here to democratize SEO and Online Marketing, making the unaffordable, affordable for every business sector, I have serviced over 5000 clients on the local market platform on a truly global scale over the past 2 years. I believe that good quality marketing should be accessible to everyone and my mission is to be at the forefront because of my client's needs.